+++ src/main/java/insane96mcp/survivalreimagined/module/world/feature/Nether.java
@@ -29,7 +29,7 @@ public class Nether extends Feature {
     @Label(name = "Disable Nether Roof and 8 block ratio", description = "Makes the nether 128 blocks high instead of 256, effectively disabling the \"Nether Roof\" and removes the 8 block ratio between nether and end.")
     public static Boolean disableNetherRoof = true;
     @Config
-    @Label(name = "Portal requires Gold Block", description = "The portal requires Gold blocks in the corners to turn it on (in the overworld).")
+    @Label(name = "Portal requires Gold Blocks or Crying Obsidians", description = "The portal requires Gold blocks or Crying Obsidians in the corners to turn it on (in the overworld).")
     public static Boolean portalRequiresGoldBlock = true;
     /*@Config
     @Label(name = "Entering the Nether requires Diamond Pickaxe advancement", description = "If the player hasn't unlocked the Diamond Pickaxe advancement he can't enter the Nether.")
@@ -62,7 +62,7 @@ public class Nether extends Feature {
         BlockPos.betweenClosed(event.getPortalSize().bottomLeft.below().relative(event.getPortalSize().rightDir.getOpposite(), 1),
                 event.getPortalSize().bottomLeft.relative(Direction.UP, event.getPortalSize().height).relative(event.getPortalSize().rightDir, event.getPortalSize().width))
                 .forEach((pos) -> {
-            if (level.getBlockState(pos).is(Blocks.GOLD_BLOCK))
+            if (level.getBlockState(pos).is(Blocks.GOLD_BLOCK) || level.getBlockState(pos).is(Blocks.CRYING_OBSIDIAN))
                 diamondBlocks.increment();
         });
         if (diamondBlocks.getValue() < 4) {
@@ -70,7 +70,7 @@ public class Nether extends Feature {
 
             for(Player player : level.players()) {
                 if (aabb.contains(player.getX(), player.getY(), player.getZ())) {
-                    player.sendSystemMessage(Component.literal("The portal needs gold blocks in the corner to be activated"));
+                    player.sendSystemMessage(Component.literal("The portal needs Gold Blocks or Crying Obsidian in the corner to be activated"));
                 }
             }
             event.setCanceled(true);