+++ changelog.md
@@ -4,6 +4,7 @@
 * Global experience multiplier increased (1.3x -> 1.5x)
 * Block experience multiplier reduced (2.25x -> 2x)
 * Overall reduced forge smashes required
+* Portal now requires either Gold Blocks or Crying obsidian
 
 ## 3.11.0
 * Added Forge